fix(frontend): update eslint-plugin-react-refresh to Vite preset
- Upgrade eslint-plugin-react-refresh from 0.4.26 to 0.5.0
- Use new Vite preset instead of manual plugin/rule configuration
- Extract useErrorHandler hook to separate file for react-refresh compliance
