+++ frontend/package.json
@@ -75,7 +75,7 @@
     "@vitest/ui": "^4.0.16",
     "eslint": "^9.31.0",
     "eslint-plugin-react-hooks": "^7.0.0",
-    "eslint-plugin-react-refresh": "^0.4.5",
+    "eslint-plugin-react-refresh": "^0.5.0",
     "jsdom": "^28.0.0",
     "pocketbase-typegen": "^1.3.3",
     "typescript": "^5.9.3",
